requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
boto3==1.35.0
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-core==0.1.0
langgraph==0.0.20
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
google-api-python-client==2.95.0